fix merge probelm

+++ include/janssonpath_error.h
@@ -5,7 +5,7 @@
 
 typedef struct jsonpath_error_t {
 	bool abort; // true if not recoverable
-	long unsigned int code; // error code. 0 for OK
+	long long unsigned int code; // error code. 0 for OK
 	const char* reason; // Null-terminated string for description text of error
 	const void* extra; // extra information. typically position of error occuring
 }jsonpath_error_t;
